requirements.txt
numpy
pandas
pandasql
google.cloud
google-cloud-bigquery
python-dotenv